test: add test for item serialization
